Move initialized and intializing tests
